+++ packages/algob/src/lib/algo-operator.ts
@@ -1,6 +1,5 @@
 import { encodeNote, parseSSCAppArgs, types as rtypes } from "@algorand-builder/runtime";
-import type { LogicSigArgs } from "algosdk";
-import algosdk from "algosdk";
+import algosdk, { LogicSigArgs } from "algosdk";
 
 import { BuilderError } from "../internal/core/errors";
 import { ERRORS } from "../internal/core/errors-list";
@@ -45,8 +44,11 @@ export interface AlgoOperator {
     txWriter: txWriter,
     scTmplParams?: SCParams) => Promise<SSCInfo>
   waitForConfirmation: (txId: string) => Promise<algosdk.ConfirmedTxInfo>
-  optInToASA: (
+  optInAcountToASA: (
     asaName: string, assetIndex: number, account: rtypes.Account, params: rtypes.TxParams
+  ) => Promise<void>
+  optInLsigToASA: (
+    asaName: string, assetIndex: number, lsig: algosdk.LogicSig, params: rtypes.TxParams
   ) => Promise<void>
   optInToASAMultiple: (
     asaName: string, asaDef: rtypes.ASADef,
@@ -108,7 +110,7 @@ export class AlgoOperatorImpl implements AlgoOperator {
     return rawSignedTxn.length;
   }
 
-  async _optInToASA (
+  async _optInAcountToASA (
     asaName: string, assetIndex: number, account: rtypes.Account, params: algosdk.SuggestedParams
   ): Promise<void> {
     console.log(`ASA ${account.name} opt-in for ASA ${asaName}`);
@@ -118,11 +120,23 @@ export class AlgoOperatorImpl implements AlgoOperator {
     await this.waitForConfirmation(txInfo.txId);
   }
 
-  async optInToASA (
+  async optInAcountToASA (
     asaName: string, assetIndex: number, account: rtypes.Account, flags: rtypes.TxParams
   ): Promise<void> {
     const txParams = await tx.mkTxParams(this.algodClient, flags);
-    await this._optInToASA(asaName, assetIndex, account, txParams);
+    await this._optInAcountToASA(asaName, assetIndex, account, txParams);
+  }
+
+  async optInLsigToASA (
+    asaName: string, assetIndex: number, lsig: algosdk.LogicSig, flags: rtypes.TxParams
+  ): Promise<void> {
+    console.log(`Contract ${lsig.address()} opt-in for ASA ${asaName}`);
+    const txParams = await tx.mkTxParams(this.algodClient, flags);
+
+    const optInLsigToASATx = tx.makeASAOptInTx(lsig.address(), assetIndex, txParams);
+    const rawLsigSignedTx = algosdk.signLogicSigTransactionObject(optInLsigToASATx, lsig).blob;
+    const txInfo = await this.algodClient.sendRawTransaction(rawLsigSignedTx).do();
+    await this.waitForConfirmation(txInfo.txId);
   }
 
   async optInToASAMultiple (
@@ -137,7 +151,7 @@ export class AlgoOperatorImpl implements AlgoOperator {
       accounts,
       flags.creator);
     for (const account of optInAccounts) {
-      await this._optInToASA(asaName, assetIndex, account, txParams);
+      await this._optInAcountToASA(asaName, assetIndex, account, txParams);
     }
   }